// src/questions/dto/question.output.dto.ts
import { AnswerOutputDTO } from '../../answers/dto/answer.output.dto';
import { MediaFileOutputDTO } from '../../mediaFiles/dto/mediaFile.output.dto';

export class QuestionOutputDTO {
	id: number;
	titleOfQuestion: string;
	explanationText: string | null;
	weight: number;
	oneAnswer: boolean;
	answerType: string | null;
	mediaFile: MediaFileOutputDTO[] | undefined;
	answers: AnswerOutputDTO[] | undefined;
}